#10 Change location of composables functions
