--- resources/meetups/2023-02-20-laravel-poland-meetup-24-test.json
@@ -1,26 +1,0 @@
-{
-  "name": "Laravel Poland Meetup #24",
-  "date": "2023-02-20 18:00:00",
-  "location": "online",
-  "organizers": ["@blumilk", "@highsolutions"],
-  "presentations": [
-    {
-      "title": "Lorem ipsum dolor sit amet",
-      "speakers": [
-        {
-          "person": "@kamilpiech",
-          "company": "@blumilk"
-        }
-      ]
-    },
-    {
-      "title": "consectetur adipiscing elit, sed do eiusmod tempor",
-      "speakers": [
-        {
-          "person": "@tomekzmudzinski",
-          "company": "@trui"
-        }
-      ]
-    }
-  ]
-}
